+++ .classpath
@@ -6,9 +6,9 @@
 		</attributes>
 	</classpathentry>
 	<classpathentry kind="src" path="src"/>
-	<classpathentry kind="lib" path="lib/cplex.jar">
+	<classpathentry kind="lib" path="libMac/cplex.jar">
 		<attributes>
-			<attribute name="org.eclipse.jdt.launching.CLASSPATH_ATTR_LIBRARY_PATH_ENTRY" value="Seminar/lib"/>
+			<attribute name="org.eclipse.jdt.launching.CLASSPATH_ATTR_LIBRARY_PATH_ENTRY" value="Seminar/libMac"/>
 		</attributes>
 	</classpathentry>
 	<classpathentry kind="output" path="bin"/>
